@@ -1,33 +1,27 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web" DefaultTargets="Publish">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
     <PublishDir>$(OutDir)_PublishedWebsites\NetCoreWebApp\</PublishDir>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Lib\**" />
     <Compile Remove="_PublishedWebsites\**" />
-    <Content Remove="Lib\**" />
     <Content Remove="_PublishedWebsites\**" />
-    <EmbeddedResource Remove="Lib\**" />
     <EmbeddedResource Remove="_PublishedWebsites\**" />
-    <None Remove="Lib\**" />
     <None Remove="_PublishedWebsites\**" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.2.0" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.4" />
+    <PackageReference Include="IGNQuery.MySql" Version="1.0.7" />
+    <PackageReference Include="IGNQuery.SqlServer" Version="1.0.9" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="5.0.0" />
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="5.0.0" />
     <PackageReference Include="MySql.Data" Version="8.0.22" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <PackageReference Include="System.Data.SqlClient" Version="4.8.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\IGNAuthentication.Data\IGNAuthentication.Data.csproj" />
+    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.8.0" />
   </ItemGroup>
 
   <ItemGroup>
@@ -37,12 +31,19 @@
     <Reference Include="IGNAuthentication.EmailValidator">
       <HintPath>..\IGNAuthentication.Data\Lib\IGNAuthentication.EmailValidator.dll</HintPath>
     </Reference>
+    <Reference Include="IGNKeyGen">
+      <HintPath>Lib\IGNKeyGen.dll</HintPath>
+    </Reference>
   </ItemGroup>
 
   <ItemGroup>
     <None Update="Programs\IGNOfflineActivator.zip">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Lib\" />
   </ItemGroup>
 
 </Project>
