@@ -1,17 +1,18 @@
 ﻿using System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using IGNAuthentication.Data;
-using IGNAuthentication.Domain.Interfaces;
+using Microsoft.Extensions.Hosting;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.IdentityModel.Tokens;
 using System.Text;
 using IGNAuthentication.Domain.Interfaces.Services;
 using IGNAuthentication.Domain.Services;
+using IGNQuery.Interfaces;
+using IGNQuery.MySql;
+using IGNQuery.SqlServer;
 
 namespace IGNLogin
 {
@@ -24,6 +25,20 @@ namespace IGNLogin
 
         public IConfiguration Configuration { get; }
 
+        private MySqlDataProvider InitMySqlDataProvider(ILogger logger,string activationEmail)
+        {
+            var mysqldp = new MySqlDataProvider(activationEmail);
+            mysqldp.queryToOutput = true;
+            return mysqldp;
+        }
+
+        private MsSqlDataProvider InitMsSqlDataProvider(ILogger logger, string activationEmail)
+        {
+            var mssqldp = new MsSqlDataProvider(activationEmail);
+            mssqldp.queryToOutput = true;
+            return mssqldp;
+        }
+
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
@@ -31,6 +46,7 @@ namespace IGNLogin
             var connectionString = Environment.GetEnvironmentVariable("MYSQL_CONNECTION_STRING");
             var adminAccessCode = Environment.GetEnvironmentVariable("ADMIN_CODE");
             var msSqlConnectionString = Environment.GetEnvironmentVariable("SQLSERVER_CONNECTION_STRING");
+            var activationMail = Environment.GetEnvironmentVariable("ACTIVATION_MAIL");
             if (string.IsNullOrWhiteSpace(connectionString) && string.IsNullOrWhiteSpace(msSqlConnectionString))
             {
             }
@@ -38,11 +54,11 @@ namespace IGNLogin
             {
                 if (string.IsNullOrWhiteSpace(msSqlConnectionString))
                 {
-                    services.AddSingleton<IDataProvider, MySqlDataProvider>(sp=>new MySqlDataProvider(sp.GetService<ILogger>()));
+                    services.AddSingleton<IDataProvider, MySqlDataProvider>(sp => InitMySqlDataProvider(sp.GetService<ILogger>(),activationMail));
                 }
                 else
                 {
-                    services.AddSingleton<IDataProvider, MsSqlDataProvider>(sp => new MsSqlDataProvider(sp.GetService<ILogger>()));
+                    services.AddSingleton<IDataProvider, MsSqlDataProvider>(sp => InitMsSqlDataProvider(sp.GetService<ILogger>(), activationMail));
                 }
             }
             services.AddScoped<IUserService, UserService>(sp => new UserService(sp.GetService<IDataProvider>()));
@@ -62,7 +78,7 @@ namespace IGNLogin
             {
                 x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
                 x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            }).AddJwtBearer(x=>
+            }).AddJwtBearer(x =>
             {
                 x.RequireHttpsMetadata = false;
                 x.SaveToken = true;
@@ -74,22 +90,30 @@ namespace IGNLogin
                     ValidIssuer = Environment.GetEnvironmentVariable("ISSUER"),
                     ValidateAudience = false
                 };
-            });            
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            });
+            services.AddControllers();
+            services.AddRazorPages();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
             app.UseCors();
-            app.UseAuthentication();
-            app.UseMvc();
             app.UseDefaultFiles();
             app.UseStaticFiles();
+            app.UseRouting();
+
+            app.UseAuthentication();
+            app.UseAuthorization();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+                endpoints.MapRazorPages();
+            });
         }
     }
 }
